chore: move error type to its own file
