+++ apps/cli/src/main.rs
@@ -1,13 +1,15 @@
 use crate::cli::Cli;
 use crate::db::init_db;
+use crate::error::CliError;
 use crate::handlers::event::handle_event;
 use crate::handlers::sync::handle_sync;
-use crate::utils::{init_logger, CliError};
+use crate::utils::init_logger;
 use clap::Parser;
 use log::error;
 
 mod cli;
 mod db;
+mod error;
 mod event;
 mod handlers;
 mod sync;
