+++ apps/cli/src/utils.rs
@@ -5,32 +5,10 @@ use std::{
 
 use common::keyring::Keyring;
 use env_logger::Builder;
-use log::{error, LevelFilter};
+use log::LevelFilter;
 use rusqlite::Connection;
-use thiserror::Error;
 
-use crate::db::migrations;
-
-#[derive(Error, Debug)]
-pub enum CliError {
-    #[error("Database error: {0}")]
-    Db(#[from] rusqlite::Error),
-
-    #[error("Common error: {0}")]
-    Common(#[from] common::error::CommonError),
-
-    #[error("I/O error: {0}")]
-    Io(#[from] std::io::Error),
-
-    #[error("Migration error: {0}")]
-    Migration(#[from] refinery::Error),
-
-    #[error("Expected {0} command, but received a different variant")]
-    VariantMismatch(String),
-
-    #[error("Serde json error: {0}")]
-    Json(#[from] serde_json::Error),
-}
+use crate::{db::migrations, error::CliError};
 
 /// Extracts the project name from the project path
 pub fn extract_project_name<T: AsRef<Path>>(project_path: T) -> String {
